+++ packages/react-components/react-headless-components-preview/library/src/components/Accordion/AccordionItem/AccordionItem.types.ts
@@ -1,13 +1,10 @@
-import type {
-  AccordionItemSlots as AccordionItemBaseSlots,
-  AccordionItemProps as AccordionItemBaseProps,
-  AccordionItemContextValues as AccordionItemBaseContextValues,
-  AccordionItemState as AccordionItemBaseState,
+import type { AccordionItemState as AccordionItemBaseState } from '@fluentui/react-accordion';
+
+export type {
+  AccordionItemProps,
+  AccordionItemSlots,
+  AccordionItemContextValues as AccordionItemContextValues,
 } from '@fluentui/react-accordion';
-
-export type AccordionItemSlots = AccordionItemBaseSlots;
-
-export type AccordionItemProps = AccordionItemBaseProps;
 
 export type AccordionItemState = AccordionItemBaseState & {
   root: {
@@ -21,5 +18,3 @@ export type AccordionItemState = AccordionItemBaseState & {
     'data-open'?: string;
   };
 };
-
-export type AccordionItemContextValues = AccordionItemBaseContextValues;