+++ packages/react-components/react-headless-components-preview/library/src/components/Breadcrumb/BreadcrumbDivider/renderBreadcrumbDivider.ts
@@ -1,6 +1,1 @@
-import { renderBreadcrumbDivider_unstable } from '@fluentui/react-breadcrumb';
-
-/**
- * Renders the final JSX of the BreadcrumbDivider component, given the state.
- */
-export const renderBreadcrumbDivider = renderBreadcrumbDivider_unstable;
+export { renderBreadcrumbDivider_unstable as renderBreadcrumbDivider } from '@fluentui/react-breadcrumb';