+++ packages/react-components/react-headless-components-preview/library/src/components/Divider/renderDivider.ts
@@ -1,6 +1,1 @@
-import { renderDivider_unstable } from '@fluentui/react-divider';
-
-/**
- * Renders the final JSX of the Divider component, given the state.
- */
-export const renderDivider = renderDivider_unstable;
+export { renderDivider_unstable as renderDivider } from '@fluentui/react-divider';