+++ packages/react-components/react-headless-components-preview/library/src/components/Drawer/DrawerHeaderTitle/renderDrawerHeaderTitle.ts
@@ -1,6 +1,1 @@
-import { renderDrawerHeaderTitle_unstable } from '@fluentui/react-drawer';
-
-/**
- * Renders the final JSX of the DrawerHeaderTitle component, given the state.
- */
-export const renderDrawerHeaderTitle = renderDrawerHeaderTitle_unstable;
+export { renderDrawerHeaderTitle_unstable as renderDrawerHeaderTitle } from '@fluentui/react-drawer';