+++ packages/react-components/react-headless-components-preview/library/src/components/Menu/Menu.types.ts
@@ -1,11 +1,17 @@
-import type { MenuContextValue } from '@fluentui/react-menu';
+import type { MenuBaseProps } from '@fluentui/react-menu';
+import type { PositioningShorthand } from '../../positioning';
+
+export type MenuProps = Omit<MenuBaseProps, 'positioning'> & {
+  /**
+   * Configures the positioned menu.
+   */
+  positioning?: PositioningShorthand;
+};
 
 export type {
-  MenuBaseProps as MenuProps,
   MenuBaseState as MenuState,
   MenuContextValue,
   MenuOpenChangeData,
   MenuOpenEvent,
+  MenuContextValues,
 } from '@fluentui/react-menu';
-
-export type MenuContextValues = { menu: MenuContextValue };