+++ scripts/beachball/base.config.js
@@ -22,6 +22,7 @@ const config = {
     '**/.storybook/**',
     '**/bundle-size/**',
     '**/monosize.config.mjs',
+    '**/bundle-isolation.config.json',
     '**/common/isConformant.ts',
     '**/src/testing/**',
     '**/src/e2e/**',
@@ -30,7 +31,6 @@ const config = {
     '**/SPEC*.md',
     '**/tests/**',
   ],
-  scope: ['!packages/fluentui/*'],
   changehint: "Run 'yarn change' to generate a change file",
 };